Move ceramic measures validation function to the component
